[ALT-DOC] INTERFACES & DATA FOR CANBUS TOPOLOGY
add and modify diagrams to better represent the data exchanged as part
of the telecommanded mode change pipeline:
- [CEI] System Interfaces, with (G/S--Op.) and (S/C--G/S) interfaces and
  data exchange items expanded to subsystem-level granularity by the G/S
-- for context, with a "smart Comms" subsystem, this expansion would
have been performed by Comms itself
- [CII] Subsystem Interfaces, delegating the (S/C--G/S) interface to the
  Comms Subsystem and introducing an (Any-non-Comms--Comms) interface
representing the internal transmission of addressed packets, allocating the same data exchange items
- change first OBC function at Physical level from "Listen for ..." to
  "Eavesdrop for ..."
- change all inter-MCU Physical Links from I2C to CAN

this infrastructure architecture is modelled for CAN bus:
- bus topology (all subsystems are daisy-chained, including comms)
- eavesdroppable (addresses are assigned to messages, not to devices)

the introduction of these diagrams is made in preparation for future
changes, to highlight the contrasts between alternative architectures.
as such, the contents of the current changes are only for documentation
purposes, and will be overwritten in the next commits to implement a
model more representative of the actual system.