[FIX] Fix Orbit Entry States
at System Level, modify [MSM] System States:
- remove Contact-Seeking state
- move ejection transition target to Detumbling
- rename Detumbling -> Deployment transition to the only occurrence of
  Ejection Loop Progress

also, in System Matrix: States vs Functions' Deployment state, add
"Receive mode change ..." and "Send mode change ..." functions and swap
"Operate nominal power" to "Operate high power".

at Logical Level, modify
- [MSM] Comms States:
	- add Detumbling and Deployment state realisation to Send state
	- add new state loop: Entry Orbit Reverse Link Loop
	  (Idle -> Send)
	- remove realisations from the forward loop transitions and
	  allocate them to the reverse loop instead
- [MSM] ADCS States:
	- add Deployment state realisation to Idle state
	- change Idle -> Actuate transition trigger description to
	  Ejection
	- add Ejection Loop Entry transition realisation to Idle ->
	  Actuate
- [MSM] EPS States:
	- change Idle -> High-Power transition trigger description to a
	  more generic High demand
	- add Ejection Loop Entry transition realistaion to Idle ->
	  High-Power
	- add Ejection Loop Exit transition realisation to High-Power ->
	  Idle

note: this region of the FSM and its implications will require more work
when actually considering the launch stage of the mission (e.g. uplink
and and downlink actions which are kept in separate states in the main
operation region are merged together in this region for simplicity);
however, refer to this changeset for an idea of the work to do when
modifying the Finite State Machine representations :)